Harden error handling and validation in ASCII tables, compression, and heap descriptors
- ASCII table field extraction now returns `Result` to surface non-UTF-8
  bytes
  as a malformed-file error instead of silently decoding corrupt data as
  blank
- RICE codec now validates `BYTEPIX ∈ {1, 2, 4}` per FITS spec,
  rejecting
  untrusted headers that would otherwise decode with mismatched masks
- Table compression `be_to_i64` treats byte columns as unsigned
  (matching image
  RICE path), with clarifying comment on sign convention
- Binary table heap offsets/counts saturate on 32-bit targets to fail
  bounds
  checks rather than wrap into spurious in-range values
- VLA descriptor write now validates 32-bit `P` columns don't overflow
  `u32::MAX`
- WCS `pixel_to_world` and `world_to_pixel` now document their panic
  preconditions
- Time leap-second table now documents clamping behavior outside
  1972–2017
- Writer `into_inner` now documents flush semantics

+++ src/compress/table.rs
@@ -384,11 +384,13 @@ fn decompress_column(bytes: &[u8], m: &ColMeta, rows: usize) -> Result<Vec<u8>> 
     Ok(cm)
 }
 
-/// Decode big-endian integers of `bytepix` bytes into `i64` values (signed),
-/// widening in a single pass (no intermediate narrowed `Vec`).
+/// Decode big-endian integers of `bytepix` bytes into `i64`, widening in a single
+/// pass. A `B` column is an *unsigned* byte (matching the image RICE path); the
+/// 16/32-bit forms are signed two's-complement. Sign is moot for the RICE round-trip
+/// (the codec masks to the pixel width) but this keeps one convention across paths.
 fn be_to_i64(bytes: &[u8], bytepix: usize) -> Vec<i64> {
     match bytepix {
-        1 => bytes.iter().map(|&b| b as i8 as i64).collect(),
+        1 => bytes.iter().map(|&b| b as i64).collect(),
         2 => bytes
             .chunks_exact(2)
             .map(|c| i16::from_be_bytes(c.try_into().unwrap()) as i64)
